Simple Math Calculator
Allows the user to enter 3 numbers into 3 separate text fields in the window. From there, the user can then click on the "Calculate" button and the "webpage" calculates addition, subtraction, multiplication, and division of the 3 numbers. In the future I will be looking to update the look and functionality of the webpage.