Simple Math Calculator
Allows the user to enter in 3 separate integers in 3 separate text fields. From there, the user can then click the calculate button and the webpage will calculate addition, subtraction, multiplication, and division of the 3 integers the user types in. In the future, I will be looking to update the look and functionality of the webpage.