Simple Recipe Page
All rights and ownership are the property of the Food Network and the associated owner's recipe. I am in no way attempting to take credit for this recipe. This is just a recipe that I was attempting to recreate as a tutorial for myself with HTML and CSS. The "Redirect Me" button in the bottom left hand corner will redirect you to the actual webpage where I found this recipe.