Simple Recipe Page Updated
This time, instead of using the table attribute in HTML in order to create this webpage, I attempted to learn how to use the Div attribute along with a CSS page. Again, I am in no way attempting to take credit for this recipe. The "Redirect Me" button at the top of the page will redirect you to the Food Network page that I used as a model to create this webpage.